--- app/concepts/data_source/file/pp/operation/add_personne_physique.rb
@@ -1,10 +1,0 @@
-module DataSource
-  module File
-    module PP
-      module Operation
-        class AddPersonnePhysique < Trailblazer::Operation
-        end
-      end
-    end
-  end
-end
