+++ app/concepts/data_source/file/pp/operation/import.rb
@@ -32,7 +32,7 @@ module DataSource
 
           def line_import_personnes_physiques(ctx, file_path:, file_reader:, **)
             file_reader.line_processing(file_path, PP_HEADER_MAPPING) do |line|
-              line_import = PP::Operation::AddPersonnePhysique.call(data: line)
+              line_import = PersonnePhysique::Operation::Create.call(data: line)
 
               return false if line_import.failure?
             end
