+++ app/concepts/personne_physique/operation/create.rb
@@ -1,0 +1,6 @@
+class PersonnePhysique
+  module Operation
+    class Create < Trailblazer::Operation
+    end
+  end
+end
