+++ spec/concepts/data_source/file/pp/operation/import_spec.rb
@@ -21,7 +21,7 @@ describe DataSource::File::PP::Operation::Import, :trb do
     it_behaves_like 'bulk import', DossierEntreprise, file, DOSSIER_ENTREPRISE_FROM_PP_HEADER_MAPPING
 
     it_behaves_like 'line import',
-      DataSource::File::PP::Operation::AddPersonnePhysique,
+      PersonnePhysique::Operation::Create,
       PersonnePhysique,
       file,
       PP_HEADER_MAPPING
